Prepare to build vRouter as Hyper-V Extension
Changes required to support Contrail on Windows.
1. Add OS-specific GUID to vr_interface
2. Use compiler-independent wrappers on __attribute__((packed))
3. Use compiler-independent wrappers on __attribute__((format))
4. Use compiler-independent wrappers on __attribute__((unused))
5. Cast pointers to uintptr_t instead of unsigned long

Partial-Bug: #1734699

+++ include/vr_queue.h
@@ -6,9 +6,11 @@
 #ifndef __VR_QUEUE_H__
 #define __VR_QUEUE_H__
 
+#include "vr_os.h"
+
 #define CONTAINER_OF(member, struct_type, pointer) \
-    ((struct_type *)((unsigned long)pointer - \
-        (size_t)&(((struct_type *)0)->member)))
+    ((struct_type *)((uintptr_t)pointer - \
+                (uintptr_t)&(((struct_type *)0)->member)))
 
 struct vr_qelem {
     struct vr_qelem *q_next;
